+++ scripts/webframeworks-deploy-tests/tests.ts
@@ -135,6 +135,14 @@ describe("webframeworks", function (this) {
               {
                 headers: [
                   {
+                    key: "x-nextjs-stale-time",
+                    value: "4294967294",
+                  },
+                  {
+                    key: "x-nextjs-prerender",
+                    value: "1",
+                  },
+                  {
                     key: "x-next-cache-tags",
                     value:
                       "_N_T_/layout,_N_T_/app/layout,_N_T_/app/image/layout,_N_T_/app/image/page,_N_T_/app/image",
@@ -143,7 +151,33 @@ describe("webframeworks", function (this) {
                 source: "/base/app/image",
               },
               {
+                source: "/base/_not-found",
                 headers: [
+                  {
+                    key: "x-nextjs-stale-time",
+                    value: "4294967294",
+                  },
+                  {
+                    key: "x-nextjs-prerender",
+                    value: "1",
+                  },
+                  {
+                    key: "x-next-cache-tags",
+                    value:
+                      "_N_T_/layout,_N_T_/_not-found/layout,_N_T_/_not-found/page,_N_T_/_not-found",
+                  },
+                ],
+              },
+              {
+                headers: [
+                  {
+                    key: "x-nextjs-stale-time",
+                    value: "4294967294",
+                  },
+                  {
+                    key: "x-nextjs-prerender",
+                    value: "1",
+                  },
                   {
                     key: "x-next-cache-tags",
                     value:
@@ -154,6 +188,14 @@ describe("webframeworks", function (this) {
               },
               {
                 headers: [
+                  {
+                    key: "x-nextjs-stale-time",
+                    value: "4294967294",
+                  },
+                  {
+                    key: "x-nextjs-prerender",
+                    value: "1",
+                  },
                   {
                     key: "x-next-cache-tags",
                     value: