+++ src/commands/appdistribution-groups-delete.ts
@@ -1,12 +1,13 @@
 import { Command } from "../command";
 import * as utils from "../utils";
 import { requireAuth } from "../requireAuth";
-import { FirebaseError } from "../error";
+import { FirebaseError, getErrMsg } from "../error";
 import { AppDistributionClient } from "../appdistribution/client";
 import { getProjectName } from "../appdistribution/options-parser-util";
 
-export const command = new Command("appdistribution:group:delete <alias>")
+export const command = new Command("appdistribution:groups:delete <alias>")
   .description("delete group from a project")
+  .alias("appdistribution:group:delete")
   .before(requireAuth)
   .action(async (alias: string, options: any) => {
     const projectName = await getProjectName(options);
@@ -14,8 +15,8 @@ export const command = new Command("appdistribution:group:delete <alias>")
     try {
       utils.logBullet(`Deleting group from project`);
       await appDistroClient.deleteGroup(`${projectName}/groups/${alias}`);
-    } catch (err: any) {
-      throw new FirebaseError(`Failed to delete group ${err}`);
+    } catch (err: unknown) {
+      throw new FirebaseError(`Failed to delete group ${getErrMsg(err)}`);
     }
     utils.logSuccess(`Group ${alias} has successfully been deleted`);
   });