+++ src/emulator/apphosting/index.ts
@@ -1,11 +1,12 @@
 import { EmulatorInfo, EmulatorInstance, Emulators } from "../types";
 import { start as apphostingStart } from "./serve";
-import { logger } from "./utils";
+import { logger } from "./developmentServer";
 interface AppHostingEmulatorArgs {
   options?: any;
   port?: number;
   host?: string;
   startCommandOverride?: string;
+  rootDirectory?: string;
 }
 
 /**
@@ -18,6 +19,7 @@ export class AppHostingEmulator implements EmulatorInstance {
   async start(): Promise<void> {
     const { hostname, port } = await apphostingStart({
       startCommand: this.args.startCommandOverride,
+      rootDirectory: this.args.rootDirectory,
     });
     this.args.options.host = hostname;
     this.args.options.port = port;