+++ src/emulator/apphosting/serve.spec.ts
@@ -4,30 +4,35 @@ import * as portUtils from "../portUtils";
 import * as spawn from "../../init/spawn";
 import * as serve from "./serve";
 import { DEFAULT_PORTS } from "../constants";
-import * as utils from "./utils";
-import * as configs from "./config";
+import * as utils from "./developmentServer";
+import * as configsImport from "./config";
+import * as projectPathImport from "../../projectPath";
+import { AppHostingYamlConfig } from "../../apphosting/yaml";
 
 describe("serve", () => {
   let checkListenableStub: sinon.SinonStub;
   let wrapSpawnStub: sinon.SinonStub;
   let spawnWithCommandStringStub: sinon.SinonStub;
-  let discoverPackageManagerStub: sinon.SinonStub;
-  let getLocalAppHostingConfigurationStub: sinon.SinonStub;
+  let detectStartCommandStub: sinon.SinonStub;
+  let configsStub: sinon.SinonStubbedInstance<typeof configsImport>;
+  let resolveProjectPathStub: sinon.SinonStub;
 
   beforeEach(() => {
     checkListenableStub = sinon.stub(portUtils, "checkListenable");
     wrapSpawnStub = sinon.stub(spawn, "wrapSpawn");
     spawnWithCommandStringStub = sinon.stub(spawn, "spawnWithCommandString");
-    discoverPackageManagerStub = sinon.stub(utils, "discoverPackageManager");
-    getLocalAppHostingConfigurationStub = sinon.stub(configs, "getLocalAppHostingConfiguration");
+    detectStartCommandStub = sinon.stub(utils, "detectStartCommand");
+    configsStub = sinon.stub(configsImport);
+    resolveProjectPathStub = sinon.stub(projectPathImport, "resolveProjectPath");
+
+    resolveProjectPathStub.returns("");
+    detectStartCommandStub.returns("npm run dev");
   });
 
   afterEach(() => {
-    checkListenableStub.restore();
     wrapSpawnStub.restore();
-    spawnWithCommandStringStub.restore();
-    discoverPackageManagerStub.restore();
-    getLocalAppHostingConfigurationStub.restore();
+    detectStartCommandStub.restore();
+    sinon.verifyAndRestore();
   });
 
   describe("start", () => {
@@ -35,7 +40,9 @@ describe("serve", () => {
       checkListenableStub.onFirstCall().returns(false);
       checkListenableStub.onSecondCall().returns(false);
       checkListenableStub.onThirdCall().returns(true);
-      getLocalAppHostingConfigurationStub.returns({ environmentVariables: {}, secrets: {} });
+      configsStub.getLocalAppHostingConfiguration.returns(
+        Promise.resolve(AppHostingYamlConfig.empty()),
+      );
       const res = await serve.start();
       expect(res.port).to.equal(DEFAULT_PORTS.apphosting + 2);
     });
@@ -43,7 +50,9 @@ describe("serve", () => {
     it("should run the custom start command if one is provided", async () => {
       const startCommand = "custom test command";
       checkListenableStub.onFirstCall().returns(true);
-      getLocalAppHostingConfigurationStub.returns({ environmentVariables: {}, secrets: {} });
+      configsStub.getLocalAppHostingConfiguration.returns(
+        Promise.resolve(AppHostingYamlConfig.empty()),
+      );
 
       await serve.start({ startCommand });