+++ src/emulator/downloadableEmulators.ts
@@ -59,20 +59,20 @@ const EMULATOR_UPDATE_DETAILS: { [s in DownloadableEmulators]: EmulatorUpdateDet
   dataconnect:
     process.platform === "darwin"
       ? {
-          version: "1.5.1",
-          expectedSize: 25289472,
-          expectedChecksum: "92c425072db66c7e2cfa40b703ed807b",
+          version: "1.7.3",
+          expectedSize: 25211648,
+          expectedChecksum: "8410794304b2ae340c3facf07d7edc16",
         }
       : process.platform === "win32"
         ? {
-            version: "1.5.1",
-            expectedSize: 25720320,
-            expectedChecksum: "2a5c654770233b740980d5f98f24be73",
+            version: "1.7.3",
+            expectedSize: 25641984,
+            expectedChecksum: "a4bd0f9d9d884528fa4494e4d7918c08",
           }
         : {
-            version: "1.5.1",
-            expectedSize: 25202840,
-            expectedChecksum: "f95156cbcac237268791638ea0eb10e7",
+            version: "1.7.3",
+            expectedSize: 25125016,
+            expectedChecksum: "48660e6370aeed973f33c3420c3255fb",
           },
 };
 
@@ -281,8 +281,8 @@ const Commands: { [s in DownloadableEmulators]: DownloadableEmulatorCommand } = 
     shell: true,
   },
   ui: {
-    binary: "node",
-    args: [getExecPath(Emulators.UI)],
+    binary: "",
+    args: [],
     optionalArgs: [],
     joinArgs: false,
     shell: false,