+++ CHANGELOG.md
@@ -1,2 +1,2 @@
-- Fix rollouts:create to handle backend regionality & other fixes. (#7862)
-- Fixed Next.js issue with PPR routes not rendering correctly. (#7625)
+- Added `--import` and `emulators:export` support to the Data Connect emulator.
+- Added `firebase.json#emulators.dataconnect.dataDir`. When set, Data Connect data will be persisted to the configured directory between emulator runs.